requirements.txt
pygame==2.5.2
azure-cognitiveservices-speech==1.37.0   # STT & TTS
openai==1.30.0                            # Azure OpenAI uses this SDK
